+++ MemoryTest/MemoryTest/HeapAllocator.cpp
@@ -1,29 +1,5 @@
 #include <cassert>
-struct  MemoryBlock
-{
-    void* pBaseAddress;
-    size_t BlockSize;
-    MemoryBlock* pNextBlock;
-};
-
-
-
-class HeapAllocator
-{
-public:
-    static HeapAllocator* Create(void* i_pHeapMemory, size_t i_HeapMemorySize, unsigned int i_NumMemBlocks);
-private:
-    HeapAllocator(void* i_pHeapMemory, size_t i_HeapMemorySize, unsigned int i_NumMemBlocks);
-
-    MemoryBlock* InitializeMemoryBlocks(void* i_pBlocksMemory, size_t i_BlocksMemorySize);
-    MemoryBlock* GetMemoryBlock();
-    void ReturnMemoryBlock(MemoryBlock* i_pFreeBlock);
-    void* alloc(size_t size);
-
-    struct MemoryBlock* pFreeMemBlocks;
-    struct MemoryBlock* pFreeList;
-    struct MemoryBlock* pOutstandingAllocations;
-};
+#include "HeapAllocator.h"
 
 
 
@@ -78,7 +54,7 @@ MemoryBlock* HeapAllocator::InitializeMemoryBlocks(void* i_pBlocksMemory, size_t
 
 MemoryBlock* HeapAllocator::GetMemoryBlock()
 {
-    assert(pFreeList != nullptr);
+    //assert(pFreeList != nullptr);
 
     MemoryBlock* pReturnBlock = pFreeMemBlocks;
     pFreeMemBlocks = pFreeMemBlocks->pNextBlock;
